modify Chunking Configuration to be with Additional Instructions component
