+++ frontend/src/components/Popups/GraphEnhancementDialog/index.tsx
@@ -9,7 +9,6 @@ import DeduplicationTab from './Deduplication';
 import { tokens } from '@neo4j-ndl/base';
 import PostProcessingCheckList from './PostProcessingCheckList';
 import AdditionalInstructionsText from './AdditionalInstructions';
-import ChunkingConfiguration from './ChunkingConfiguration';
 
 export default function GraphEnhancementDialog({ open, onClose }: { open: boolean; onClose: () => void }) {
   const { breakpoints } = tokens;
@@ -104,14 +103,6 @@ export default function GraphEnhancementDialog({ open, onClose }: { open: boolea
                   >
                     Post Processing Jobs
                   </Tabs.Tab>
-                  <Tabs.Tab
-                    tabId={5}
-                    htmlAttributes={{
-                      'aria-label': 'Chunking Configuration',
-                    }}
-                  >
-                    Chunking Configuration
-                  </Tabs.Tab>
                 </Tabs>
               </Flex>
             </div>
@@ -142,9 +133,6 @@ export default function GraphEnhancementDialog({ open, onClose }: { open: boolea
         </Tabs.TabPanel>
         <Tabs.TabPanel className='n-flex n-flex-col n-gap-token-4 n-p-token-6' value={activeTab} tabId={4}>
           <PostProcessingCheckList />
-        </Tabs.TabPanel>
-        <Tabs.TabPanel className='n-flex n-flex-col n-gap-token-4 n-p-token-6' value={activeTab} tabId={5}>
-          <ChunkingConfiguration/>
         </Tabs.TabPanel>
       </Dialog.Content>
     </Dialog>
